Testss

+++ src/main/java/Base/TestBase.java
@@ -1,5 +1,6 @@
 package Base;
 
+import Pages.LoginPage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import java.io.File;
@@ -10,8 +11,9 @@ import java.util.Properties;
 
 public class TestBase {
 
-    public static WebDriver driver;
+     public static WebDriver driver;
      public static Properties prop;
+     public static LoginPage login;
    static private final String filePath = System.getProperty("user.dir")+"/src/main/resources/config/config.properties";
 
 
@@ -29,13 +31,24 @@ public class TestBase {
        public void initialization(){
         //System.setProperty("webDriver.chrome.driver", System.getProperty("user.dir")+"src/drivers/chromedriver.exe");
         driver = new ChromeDriver();
+        driver.get(prop.getProperty("url"));
         driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
         driver.manage().window().fullscreen();
-        driver.get(prop.getProperty("url"));
+
+    }
+
+    public void login(){
+        login = new LoginPage(driver);
+        login.enterUserName(prop.getProperty("username"));
+        login.enterPassword(prop.getProperty("password"));
+        login.clickLoginButton();
+
     }
 
     public void tearDown() {
-        driver.manage().deleteAllCookies();
-        driver.quit();
+        if(driver != null) {
+            driver.manage().deleteAllCookies();
+            driver.quit();
+        }
     }
 }
